refactor: extract serve subcommand and add CLI subcommand router
Move daemon startup logic from main() into runServe() in serve.go,
refactor main.go to a subcommand router (serve/get/reconcile/suspend/
resume/logs/stats/version) with full backward compatibility for bare
flags, and add stub files plus version.go and output.go scaffolding
for future CLI commands.
