+++ cmd/dockcd/get.go
@@ -1,0 +1,8 @@
+package main
+
+import "fmt"
+
+func runGet(args []string) int {
+	fmt.Println("not implemented: get")
+	return 1
+}
