+++ cmd/dockcd/logs.go
@@ -1,0 +1,8 @@
+package main
+
+import "fmt"
+
+func runLogs(args []string) int {
+	fmt.Println("not implemented: logs")
+	return 1
+}
