+++ cmd/dockcd/reconcile.go
@@ -1,0 +1,8 @@
+package main
+
+import "fmt"
+
+func runReconcile(args []string) int {
+	fmt.Println("not implemented: reconcile")
+	return 1
+}
