+++ cmd/dockcd/resume.go
@@ -1,0 +1,8 @@
+package main
+
+import "fmt"
+
+func runResume(args []string) int {
+	fmt.Println("not implemented: resume")
+	return 1
+}
