+++ cmd/dockcd/stats.go
@@ -1,0 +1,8 @@
+package main
+
+import "fmt"
+
+func runStats(args []string) int {
+	fmt.Println("not implemented: stats")
+	return 1
+}
