+++ cmd/dockcd/suspend.go
@@ -1,0 +1,8 @@
+package main
+
+import "fmt"
+
+func runSuspend(args []string) int {
+	fmt.Println("not implemented: suspend")
+	return 1
+}
